# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BankManager API",
        "description": "API de gestion bancaire pour BankManager",
        "contact": {
            "email": "contact@bankmanager.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister tous les comptes",
                "description": "Récupère la liste paginée de tous les comptes bancaires",
                "operationId": "2751c57512eeb88bdfc1b23232202bd4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "client_id",
                        "in": "query",
                        "description": "Filtrer par client UUID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de compte",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "epargne",
                                "cheque"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "bloque",
                                "ferme"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "read-comptes"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte",
                "description": "Crée un nouveau compte bancaire pour un client existant",
                "operationId": "86ebad3737ad4fc8e19637ac3b02a34c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "type"
                                ],
                                "properties": {
                                    "client_id": {
                                        "description": "UUID du client",
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "type": {
                                        "description": "Type de compte",
                                        "type": "string",
                                        "enum": [
                                            "epargne",
                                            "cheque"
                                        ]
                                    },
                                    "solde_initial": {
                                        "description": "Solde initial",
                                        "type": "number",
                                        "format": "decimal",
                                        "default": 0
                                    },
                                    "devise": {
                                        "description": "Devise",
                                        "type": "string",
                                        "default": "XOF",
                                        "maxLength": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Compte"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Client non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-comptes"
                        ]
                    }
                ]
            }
        },
        "/api/comptes/{id}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Afficher un compte",
                "description": "Récupère les détails d'un compte bancaire spécifique",
                "operationId": "3c25514451451a34a95f069a81ec4d81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Compte"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "read-comptes"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Mettre à jour un compte",
                "description": "Met à jour les informations d'un compte bancaire",
                "operationId": "5a8a43ecdd5b85083835865ccb3db211",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "statut": {
                                        "description": "Nouveau statut du compte",
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "bloque",
                                            "ferme"
                                        ]
                                    },
                                    "notes": {
                                        "description": "Notes sur le compte",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Compte"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-comptes"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Fermer un compte",
                "description": "Ferme un compte bancaire (soft delete)",
                "operationId": "8d0ba54e8fa37c4210d09e91a255f78e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Compte fermé avec succès"
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-comptes"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Compte": {
                "title": "Compte",
                "description": "Modèle représentant un compte bancaire",
                "properties": {
                    "id": {
                        "description": "UUID du compte",
                        "type": "string",
                        "format": "uuid"
                    },
                    "numero_compte": {
                        "description": "Numéro du compte",
                        "type": "string",
                        "maxLength": 20
                    },
                    "titulaire": {
                        "description": "Nom du titulaire",
                        "type": "string",
                        "maxLength": 200
                    },
                    "type": {
                        "description": "Type de compte",
                        "type": "string",
                        "enum": [
                            "epargne",
                            "cheque"
                        ]
                    },
                    "solde": {
                        "description": "Solde du compte",
                        "type": "number",
                        "format": "decimal"
                    },
                    "solde_formate": {
                        "description": "Solde formaté avec devise",
                        "type": "string"
                    },
                    "devise": {
                        "description": "Devise du compte",
                        "type": "string",
                        "maxLength": 3
                    },
                    "date_creation": {
                        "description": "Date de création du compte",
                        "type": "string",
                        "format": "date"
                    },
                    "statut": {
                        "description": "Statut du compte",
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "ferme"
                        ]
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "datetime"
                    },
                    "client": {
                        "$ref": "#/components/schemas/Client"
                    },
                    "transactions": {
                        "description": "Transactions du compte",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Transaction"
                        }
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "title": "Transaction",
                "description": "Modèle représentant une transaction bancaire",
                "properties": {
                    "id": {
                        "description": "UUID de la transaction",
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "description": "Type de transaction",
                        "type": "string",
                        "enum": [
                            "depot",
                            "retrait",
                            "virement_entrant",
                            "virement_sortant",
                            "frais",
                            "interet"
                        ]
                    },
                    "type_libelle": {
                        "description": "Libellé du type de transaction",
                        "type": "string"
                    },
                    "montant": {
                        "description": "Montant de la transaction",
                        "type": "number",
                        "format": "decimal"
                    },
                    "montant_formate": {
                        "description": "Montant formaté avec devise",
                        "type": "string"
                    },
                    "devise": {
                        "description": "Devise de la transaction",
                        "type": "string",
                        "maxLength": 3
                    },
                    "description": {
                        "description": "Description de la transaction",
                        "type": "string",
                        "nullable": true
                    },
                    "date_transaction": {
                        "description": "Date et heure de la transaction",
                        "type": "string",
                        "format": "datetime"
                    },
                    "statut": {
                        "description": "Statut de la transaction",
                        "type": "string",
                        "enum": [
                            "en_attente",
                            "validee",
                            "annulee",
                            "echouee"
                        ]
                    },
                    "compte_destinataire": {
                        "description": "Numéro du compte destinataire",
                        "type": "string",
                        "nullable": true
                    },
                    "nom_destinataire": {
                        "description": "Nom du destinataire",
                        "type": "string",
                        "nullable": true
                    },
                    "frais": {
                        "description": "Frais de transaction",
                        "type": "number",
                        "format": "decimal"
                    },
                    "solde_apres": {
                        "description": "Solde après transaction",
                        "type": "number",
                        "format": "decimal"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "datetime"
                    },
                    "compte": {
                        "$ref": "#/components/schemas/Compte"
                    }
                },
                "type": "object"
            },
...
            "PaginationLinks": {
                "title": "PaginationLinks",
                "description": "Liens de pagination",
                "properties": {
                    "first": {
                        "description": "Lien vers la première page",
                        "type": "string"
                    },
                    "last": {
                        "description": "Lien vers la dernière page",
                        "type": "string"
                    },
                    "prev": {
                        "description": "Lien vers la page précédente",
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "description": "Lien vers la page suivante",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "PaginationMeta",
                "description": "Métadonnées de pagination",
                "properties": {
                    "current_page": {
                        "description": "Page actuelle",
                        "type": "integer"
                    },
                    "from": {
                        "description": "Premier élément de la page",
                        "type": "integer"
                    },
                    "last_page": {
                        "description": "Dernière page",
                        "type": "integer"
                    },
                    "links": {
                        "description": "Liens de pagination",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "path": {
                        "description": "URL de base",
                        "type": "string"
                    },
                    "per_page": {
                        "description": "Éléments par page",
                        "type": "integer"
                    },
                    "to": {
                        "description": "Dernier élément de la page",
                        "type": "integer"
                    },
                    "total": {
                        "description": "Nombre total d'éléments",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel Passport OAuth2 security.",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/oauth/token/refresh",
                        "scopes": {
                            "read-clients": "Lire les clients",
                            "write-clients": "Créer/modifier les clients",
                            "read-comptes": "Lire les comptes",
                            "write-comptes": "Créer/modifier les comptes",
                            "read-transactions": "Lire les transactions",
                            "write-transactions": "Créer des transactions"
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        {
            "passport": [
                "read-clients",
                "write-clients",
                "read-comptes",
                "write-comptes",
                "read-transactions",
                "write-transactions"
            ]
        }
    ]
}